more changes

+++ onnx_light/onnx_proto/onnx.cc
@@ -662,7 +662,7 @@ void TensorProto::ParseFromStream(utils::BinaryStream &stream, ParseOptions &opt
   READ_FIELD(options, stream, name)                        //
   READ_FIELD(options, stream, doc_string)                  //
   else if (static_cast<int>(field_number.field_number) == order_raw_data()) {
-    read_field_limit_parallel_nc(stream, field_number.wire_type, raw_data_, "raw_data", options);
+    read_field_limit_parallel_nc(stream, static_cast<int>(field_number.wire_type), raw_data_, "raw_data", options);
   } //
   READ_REPEATED_FIELD(options, stream, external_data)  //
   READ_REPEATED_FIELD(options, stream, metadata_props) //
